requirements.txt
openai>=1.0.0
google-generativeai>=0.3.2
anthropic>=0.3.12
gradio>=4.29.0
python-dotenv>=1.0.1
ipython>=8.0.0
requests>=2.31.0
